Make it possible to remove broken taxonomies
Due to a bug in previous versions users may have taxonomies without
IDs. This can break things but they cannot be removed as attempting to
delete the taxonomy causes a “Cheating” error to be generated.

This update removes the error and instead cycles all the taxonomies to
remove any that do not have an ID set.

addresses #19