Flatten details array to pass values to wp_kses instead of arrays
We were passing an array to wp_kses, which only accepts a string.

Passing an array to wp_kses was never officially supported and now
breaks in WP 5.3.1. https://core.trac.wordpress.org/ticket/48955

array_walk instead of array_map because we need to retain property
details strings as keys.

santize_detail callback instead of an anonymous function because we
need to use `$this->allowed_tags` in the function, and `$this` is only
supported inside anonymous functions in PHP 5.4+. (We have no stated
minimum PHP version for this plugin that I could see.)